+++ lib/cached_image_manager.dart
@@ -2,6 +2,14 @@ import 'dart:io';
 import 'dart:typed_data';
 
 abstract class CachedImageManager {
+  Future<bool> isExists(String key);
+
+  Future<bool> removeFile(String key);
+
+  Future<void> clearCache();
+
+  Future<void> dispose();
+
   Future<File> cacheBase64(
     String key,
     String base64, {